Add files via upload
In this project, I worked with an interesting dataset that I initially loaded into Jupyter Notebook for exploration. I started with Exploratory Data Analysis (EDA) to understand the structure, trends, and patterns in the data. After that, I focused on data cleaning, where I handled missing and null values carefully to ensure the dataset was reliable for further use. I also performed preprocessing on the date and time columns to make them more useful for analysis. This step-by-step process not only helped me prepare the dataset but also improved its overall quality for meaningful insights